training and development
recreated the bugdet watcher with a more modularized internal design trying to practise mvvm, unit tests etc.
the ui design is rather basic compared to before, but easier to see the workflow through intuition.

todo:
messagebox calls must be moved to views to make it more mvvm, cause it breaks the pattern.  hadn't had the time yet and i am undecided whether i should use a dialog service (as far as i understand it would be a cleaner solution, but it bloats the project) or events (which would make it mvvm, but introduces potential memory issues, if events are not unregistered).